refactor(reposicao): extrai ItensTab de AdminReposicaoPromocaoDetail
Segundo corte do god-component #1. Extrai a aba "Itens da campanha"
(tabela com edição inline + add-item form) pra ItensTab.tsx.

Boundary escolhido pra preservar comportamento exato:
- Estado de adicionar-item (addingItem, novo*, savingNovoItem) +
  handleAddItem são 100% locais à aba → movidos pra dentro do ItensTab.
  handleAddItem mantém a mesma lógica (insert + RPC expandir + invalidate
  com a MESMA query key ["promocao-itens", campanhaId]).
- Mutations de linha (update/delete) ficam no parent (são compartilhadas
  com a query) e chegam via callbacks onUpdateItem/onDeleteItem — zero
  mudança de data-flow.

Parent: remove o estado/handler de add-item + a tabela JSX, troca pelo
<ItensTab/>. Limpa 8 imports órfãos (Table*, Trash2, e as imports das
células que agora só o ItensTab usa).

ItensTab entra no strict (332). Página continua reduzindo (orquestração).

Verificado: typecheck:strict 0, bunx tsc 0, eslint 0 unused, build OK,
358 testes passando.
